+++ src/index.ts
@@ -1,4 +1,3 @@
 export { default as stricter } from './stricter';
 export { default as cli } from './cli';
-export { default as readFilesData } from './file-processor';
-export { externalUseReadDependencies as readDependencies } from './file-processor/external';
+export { parseDependencies } from './external';